% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{computeage}
\alias{computeage}
\title{Compute age. text input and output}
\usage{
computeage(textdate0, textdate1)
}
\arguments{
\item{textdate0}{character vector of dates in format YYYYMMDD.}

\item{textdate1}{character vector of dates in format YYYYMMDD.}
}
\value{
character vector of integer ages (age at last birthday)
}
\description{
Compute age. text input and output
}
