% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{computeage_dep}
\alias{computeage_dep}
\title{Compute age. text input and output}
\usage{
computeage_dep(textdate0, textdate1)
}
\arguments{
\item{textdate0}{length-1 character vector of dates in format YYYYMMDD}

\item{textdate1}{length-1 character vector of dates in format YYYYMMDD}
}
\value{
character vector of integer ages (age at last birthday)
}
\description{
Compute age. text input and output
}
