% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{csvdircheck}
\alias{csvdircheck}
\title{Check form/existence of csvdir}
\usage{
csvdircheck(csvdir, verbose = 0)
}
\arguments{
\item{csvdir}{proposed existing directory to write to (character vector of length 1)}

\item{verbose}{level of output to console}
}
\value{
either NULL or character string of existing directory to write to
}
\description{
Check form/existence of csvdir
}
