% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{my_find_all}
\alias{my_find_all}
\title{Return the all nodes found with the associated tag}
\usage{
my_find_all(node, tag)
}
\arguments{
\item{node}{xml node}

\item{tag}{attribute tag to find (character)}
}
\value{
all nodes identified by tag
}
\description{
Return the all nodes found with the associated tag
}
