% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{my_find_first}
\alias{my_find_first}
\title{Return the first node found with the associated tag}
\usage{
my_find_first(node, tag)
}
\arguments{
\item{node}{xml node.}

\item{tag}{attribute tag to find (character).}
}
\value{
first node identified by tag
}
\description{
Return the first node found with the associated tag
}
