% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{text_of_first}
\alias{text_of_first}
\title{Return the text value associated with a tag}
\usage{
text_of_first(node, tag)
}
\arguments{
\item{node}{xml node}

\item{tag}{attribute tag to find}
}
\value{
length-1 character vector containing text of tag. Value is NA
    if node is missing or if tag is missing.  Value is "" if tag is found
    for node but the tag has no value.
}
\description{
Return the text value associated with a tag
}
