% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_dir_explode.R
\name{xml_dir_explode}
\alias{xml_dir_explode}
\title{Find root node for each xml file in a directory}
\usage{
xml_dir_explode(xmldir, recursive = FALSE, verbose = 0)
}
\arguments{
\item{xmldir}{(character) Name of directory containing xml files from FORUM}

\item{recursive}{(logical) Should sub directories be included? Default is FALSE}

\item{verbose}{(integer) Amount of output to the console.  0 is none. Higher values may produce more output. Default 0.}
}
\value{
list of root nodes, one for each xml file in the directory xmldir
}
\description{
Find root node for each xml file in a directory
}
\examples{
xml_dir_explode(".")
}
