% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_dir_extract_points.R
\name{xml_dir_extract_points}
\alias{xml_dir_extract_points}
\title{Extract visual field values from all xml files in a directory}
\usage{
xml_dir_extract_points(
  xmldir,
  recursive = FALSE,
  csvdir = NULL,
  excludenonhvf = FALSE,
  eyeformat = "OD",
  dropXY = eyeformat \%in\% c("OD", "OS"),
  extra4fields = FALSE,
  verbose = 0L
)
}
\arguments{
\item{xmldir}{(character) Name of directory containing xml files from FORUM.}

\item{recursive}{(logical) Should sub directories be included? Default is
\code{FALSE}.}

\item{csvdir}{(character) Name of directory to write csv file(s) to. Default (\code{NULL}) is not to write files.}

\item{excludenonhvf}{(logical) Should files that don't have a recognizable HVF
Test Pattern be excluded from the output value/file? If \code{FALSE} (default),
then every xml file will be represented in one of the list components (returned
and/or written to files). If \code{TRUE}, Test Patterns <NA> and "" will not
be returned and/or written.}

\item{eyeformat}{See `xml_extract_points()`. Leave as \code{"OD"} (default).}

\item{dropXY}{See `xml_extract_points()`.  Leave as \code{TRUE} (default).}

\item{extra4fields}{See `xml_extract_points()`.  Leave as \code{FALSE} (default).}

\item{verbose}{(integer) Amount of output to the console.  0 is none. Higher values may produce more output. Default 0.}
}
\value{
list of character matrices. Each component is a matrix for a Test Pattern. Each row of a matrix is for one file.  First column is file name with complete path. Subsequent columns are extracted values from xml files.
}
\description{
Extract visual field values from all xml files in a directory
}
\examples{
# xml_dir_extract_points(".") # files in current directory
}
