% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_extract_points.R
\name{xml_point}
\alias{xml_point}
\title{Extract data from XML node representing HVF point}
\usage{
xml_point(pointroot, extra4fields = FALSE)
}
\arguments{
\item{pointroot}{An xml node containing information about one point in a Humphrey Visual Field}

\item{extra4fields}{Most points have 7 pieces of information that are extracted. There are 4 other fields that rarely occur.  Should these extra 4 fields (Stimulus result: SEEN vs NOT SEEN, twice; Sensitivity value, and Generalized Defect Corrected Sensitivity Deviation Flag) be extracted also? Default is \code{FALSE}.}
}
\value{
A named character vector of length 7 (or 11) with names "X" (x-coordinate), "Y" (y-coordinate), "Sens Val" (sensitivity value), "ACSDV" (age corrected sensitivity deviation value), "GDCSDV" (generalized defect corrected sensitivity deviation value), "ACSDPV" (age corrected sensitivity deviation probability value), and "GDCSDPV" (generalized defect corrected sensitivity deviation probability value). Optionally: "Stim Res" (Stimulus result: SEEN vs NOT SEEN), "Stim Res 2" (Stimulus result: SEEN vs NOT SEEN), "Sens Val 2" (Sensitivity value), and "GDCSDF" (Generalized Defect Corrected Sensitivity Deviation Flag).
}
\description{
Extract data from XML node representing HVF point
}
