% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_24dash2.R
\name{xml_points24dash2}
\alias{xml_points24dash2}
\title{Extract point information from FORUM XML file contain data from a 24-2 test}
\usage{
xml_points24dash2(
  top,
  asvector = TRUE,
  eyeformat = "OD",
  dropXY = eyeformat \%in\% c("OD", "OS"),
  verbose = 0
)
}
\arguments{
\item{top}{Root node of an xml file}

\item{asvector}{Determine whether the output is returned as a vector (TRUE, default) or as a matrix (FALSE).}

\item{eyeformat}{Determine the ordering of the response vector/matrix. If OD (default), the responses are ordered from top left to bottom right as they would appear on the paper printout for a right eye. If the xml file corresponds to a right eye, no transformation is done. If the xml file corresponds to a left eye, the locations are flipped across the y-axis. If eyeformat is OS, the responses are ordered from top left to bottom right as they would appear on the paper printout for left eye. If eyeformat is "file", the responses are ordered top left to bottom right without reference to eye. If eyeformat is "raw", the responses are in the order in which they were stored in the xml file.}

\item{dropXY}{Should the values of X and Y be included in the output for each point. The values of X and Y are standard for a given test pattern and eye laterality. For eyeformats OD and OS, the default is to drop X and Y. For eyeformats file and raw, the default is to keep X and Y.}

\item{verbose}{Control level of printing to console. The default (0) is no printing; larger values may produce more printing.}
}
\value{
a character vector with about 60 components extracted from the XML
}
\description{
Extract point information from FORUM XML file contain data from a 24-2 test
}
\examples{
   exdatadir <- system.file('extdata', package = 'FORUMVF')
   parsed <- xml2::read_xml(sprintf("\%s/testdata.xml", exdatadir))
   root <- xml2::xml_root(parsed)
   xml_points24dash2(root)
}
