% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_dir_extract_full.R
\name{xml_dir_extract_full}
\alias{xml_dir_extract_full}
\title{Extract visual field summaries and point value from all xml files in a directory}
\usage{
xml_dir_extract_full(
  xmldir,
  recursive = FALSE,
  csvdir = NULL,
  comment = FALSE,
  eyeformat = "OD",
  dropXY = eyeformat \%in\% c("OD", "OS"),
  extra4fields = FALSE,
  verbose = 0
)
}
\arguments{
\item{xmldir}{(character) Name of directory containing xml files from FORUM}

\item{recursive}{(logical) Should sub directories be included? Default is FALSE}

\item{csvdir}{(character) Name of directory to write csv file(s) to. Default (\code{NULL}) is not to write files.}

\item{comment}{(logical) Should text comment and patient comment be extracted? Default is FALSE.}

\item{eyeformat}{See `xml_extract_points()`. Leave as \code{"OD"} (default).}

\item{dropXY}{See `xml_extract_points()`.  Leave as \code{TRUE} (default).}

\item{extra4fields}{See `xml_extract_points()`.  Leave as \code{FALSE} (default).}

\item{verbose}{(integer) Amount of output to the console.  0 is none. Higher values may produce more output. Default 0.}
}
\value{
list with 3 components (invisibly). Component 1 ("main") is matrix of HVF summary data. Component 2 ("points") is a list of character matrices. Each subcomponent is a matrix for a Test Pattern containing the points data. Component 3 ("full") is the same as component 2 but includes both summary and points data.
}
\description{
Extract visual field summaries and point value from all xml files in a directory
}
\examples{
xml_dir_extract_full(".")
}
