% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_extract.R
\name{xml_extract}
\alias{xml_extract}
\title{Extract test level information from FORUM XML file}
\usage{
xml_extract(top, comments = FALSE)
}
\arguments{
\item{top}{Root node of an xml file}

\item{comments}{Should the patient and text comments be extracted? Default is FALSE.}
}
\value{
a character vector with about 60 components extracted from the XML
}
\description{
Extract test level information from FORUM XML file
}
\examples{
   # parsed <- xml2::read_xml("abc.xml")
   # root <- xml2::xml_root(parsed)
   # xml_extract(root)
}
